pyproject.toml
[tool.poetry]
name = "command-line-trader"
version = "0.1.0"
description = ""
authors = ["Travis Cammack <traviscammack@protonmail.com>"]
readme = "README.md"
packages = [{include = "clt"}]

[tool.poetry.scripts]
clt = "clt.main:cli"

[tool.poetry.dependencies]
python = "^3.10"
httpx = "^0.24.0"
pydantic = "^1.10.7"
pandas = "^2.0.0"
click = "^8.1.3"
tabulate = "^0.9.0"
pyyaml = "^6.0"
tiingo = "^0.14.0"
exchange-calendars = "^4.2.6"
black = "^23.3.0"
isort = "^5.12.0"

[tool.black]
line-length = 100

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
